requirements.txt
streamlit
gradio_client
Pillow
